requirements.txt
asgiref==3.3.1
Django==3.1.7
django-crispy-forms==1.11.1
django-filter==2.4.0
djangorestframework==3.12.2
Pillow==8.1.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
